requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.3.0=py39haa95532_0
blas=1.0=mkl
bottleneck=1.3.5=py39h080aedc_0
ca-certificates=2023.01.10=haa95532_0
et_xmlfile=1.1.0=py39haa95532_0
intel-openmp=2023.1.0=h59b6b97_46319
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
numexpr=2.8.4=py39h7b80656_1
numpy=1.26.0=py39h055cbcc_0
numpy-base=1.26.0=py39h65a83cf_0
openpyxl=3.0.10=py39h2bbff1b_0
openssl=3.0.11=h2bbff1b_2
pandas=2.0.3=py39h4ed8f06_0
pip=23.2.1=py39haa95532_0
python=3.9.18=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py39haa95532_0
setuptools=68.0.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
tzdata=2023c=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py39haa95532_0
xlsxwriter=3.1.6=pypi_0
